requirements.txt
fastapi
uvicorn
requests
sentence-transformers
google-cloud-aiplatform
python-dotenv
Datapoint
google-cloud-firestore